pyproject.toml
[tool.poetry]
name = "calibre-api"
version = "0.1.0"
description = "API per libreria Calibre con autenticazione e caching"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.12"
pyjwt = "^2.7.0"
redis = "^4.5.4"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pydantic = "^1.10.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.3.0"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
